feat: Introduce new UI components for data display, forms, and features like bidding and scoring, alongside backend updates and removal of drawer components.

+++ app/Http/Controllers/ReportController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Http\Request;
 use App\Http\Response;
 use Btx\File\Directory;
@@ -9,43 +10,51 @@ use App\Models\DynamicModel;
 use Illuminate\Support\Facades\Schema;
 use Exception;
 use Illuminate\Support\Facades\File;
+use App\Objects\DynamicCollectionExport;
+use Maatwebsite\Excel\Facades\Excel;
+use Carbon\Carbon;
+use App\Models\Setting;
 
 class ReportController extends BaseController
 {
     private $_query = null;
     private $_file = [];
     private $_columns = [];
-    public function __construct(){
+    public function __construct()
+    {
         $this->setModule('report.report');
         $this->setGridProperties([
             'filterDateRange' => true,
             'filterDateName' => 'nama',
             'multipleSelect' => false
         ]);
-        $this->setFilterColumnsLike(['masters.nama'],request('q')??'');
+        $this->setFilterColumnsLike(['masters.nama'], request('q') ?? '');
     }
 
-    public function queries(Request $request){
-       
+    public function queries(Request $request)
+    {
+
         return Response::ok('Loaded', $this->getQueryFiles());
     }
 
-    public function grid(Request $request){
+    public function grid(Request $request)
+    {
         $this->allowAccessModule($this->_module, 'view');
-        $this->_gridProperties['filterDateRange'] = $this->_gridProperties['filterDateRange']??false;
-        $this->_gridProperties['advanceFilter'] = $this->_gridProperties['advanceFilter']??true;
-        $this->_gridProperties['simpleFilter'] = $this->_gridProperties['simpleFilter']??true;
+        $this->_gridProperties['filterDateRange'] = $this->_gridProperties['filterDateRange'] ?? false;
+        $this->_gridProperties['advanceFilter'] = $this->_gridProperties['advanceFilter'] ?? true;
+        $this->_gridProperties['simpleFilter'] = $this->_gridProperties['simpleFilter'] ?? true;
         $this->_gridProperties['multipleSelect'] = false;
         $this->_gridProperties['contextMenu'] = $this->_contextMenus;
 
-        if(isset($request->path)){
-            $file = $this->getResourceQuery(trim($request->path));
-            $this->_file = $file;
-            $this->_columns = $file['columns'];
-            $this->_query = $file['query'];
+        if (isset($request->path)) {
+            $json = $this->getResourceQuery(trim($request->path));
+            $pathSql = trim(str_replace('.json', '.sql', $request->path));
+            $sql = $this->getResourceSql($pathSql);
+            $this->_file = $json;
+            $this->_columns = $json['columns'];
+            $this->_query = $sql;
         }
-
-        if($this->_query == null){
+        if ($this->_query == null) {
             return Response::ok('Loaded', [
                 'rows' => [],
                 'total' => 0,
@@ -53,7 +62,7 @@ class ReportController extends BaseController
                 'properties' => $this->_gridProperties
             ]);
         }
-        
+
         $model = (new DynamicModel())->setTable(DB::raw("({$this->_query}) as t"));
         $qtotal = clone $model->newQuery()->filter(false);
         $qrows = $model->newQuery()->filter();
@@ -66,11 +75,14 @@ class ReportController extends BaseController
             'total' => $total,
             'columns' => $this->_columns,
             'properties' => $this->_gridProperties,
-            'query' => $file
+            'query' => [
+                'query' => $this->_query
+            ]
         ]);
     }
 
-    public function getSchemas(Request $request){
+    public function getSchemas(Request $request)
+    {
         $result = [];
         $connection = config('database.default');
         $tables = [
@@ -116,6 +128,23 @@ class ReportController extends BaseController
                     $nullable = $colInfo->notnull == 0;
                     $default  = $colInfo->dflt_value;
                     $length   = null; // SQLite gak punya panjang kolom eksplisit
+                } elseif ($connection === 'pgsql') {
+                    // PostgreSQL: gunakan INFORMATION_SCHEMA
+                    $info = DB::select("
+                        SELECT 
+                            column_name, 
+                            column_default, 
+                            is_nullable, 
+                            character_maximum_length
+                        FROM information_schema.columns
+                        WHERE table_name = ? AND table_schema = 'public'
+                    ", [$table]);
+
+                    $colInfo = collect($info)->firstWhere('column_name', $column);
+
+                    $nullable = $colInfo->is_nullable === 'YES';
+                    $default  = $colInfo->column_default;
+                    $length   = $colInfo->character_maximum_length;
                 } elseif ($connection === 'mysql') {
                     // MySQL: gunakan INFORMATION_SCHEMA
                     $info = DB::select("
@@ -146,25 +175,26 @@ class ReportController extends BaseController
         return Response::ok('Loaded', $result);
     }
 
-    public function preview(Request $request){
+    public function preview(Request $request)
+    {
         $this->allowAccessModule('report.builder', 'view');
         $q = trim($request->rawQuery);
-        if(!$this->protectQuery($q)) return Response::badRequest('Query ini tidak dapat diproses');
-        $query = $q. " LIMIT {$request->_limit} OFFSET {$request->_page}";
+        if (!$this->protectQuery($q)) return Response::badRequest('Query ini tidak dapat diproses');
+        $query = $q . " LIMIT {$request->_limit} OFFSET {$request->_page}";
         $rows = DB::select($query);
         $total = DB::select("SELECT COUNT(*) as count FROM ({$q}) as sub")[0]->count;
-        if($total > 0){
+        if ($total > 0) {
             $rawcolumns = array_keys((array)$rows[0]);
         } else {
             $rawcolumns = [];
         }
         $columns = [];
 
-        foreach($rawcolumns as $c){
+        foreach ($rawcolumns as $c) {
             array_push($columns, [
                 "name"          => $c,
                 "required"      => true,
-                "label"         => str_replace('_',' ',strtoupper($c)),
+                "label"         => str_replace('_', ' ', strtoupper($c)),
                 "align"         => "left",
                 "field"         => $c,
                 "sortable"      => true,
@@ -182,7 +212,8 @@ class ReportController extends BaseController
         ]);
     }
 
-    public function saveQuery(Request $request){
+    public function saveQuery(Request $request)
+    {
         $this->allowAccessModule('report.builder', 'create');
 
         $rules = [
@@ -192,12 +223,12 @@ class ReportController extends BaseController
         $data = $this->validate($rules);
         if ($data instanceof \Illuminate\Http\JsonResponse) return $data;
 
-        if(!$this->protectQuery($data['query'])) return Response::badRequest('Query ini tidak dapat diproses');
+        if (!$this->protectQuery($data['query'])) return Response::badRequest('Query ini tidak dapat diproses');
         try {
             $columns = [];
             $types = [];
-            $rows = DB::select($data['query']." limit 1");
-            if(count($rows) > 0){
+            $rows = DB::select($data['query'] . " limit 1");
+            if (count($rows) > 0) {
                 $rawcolumns = array_keys((array)$rows[0]);
                 foreach ($rows[0] as $key => $value) {
                     $types[$key] = gettype($value);
@@ -205,12 +236,12 @@ class ReportController extends BaseController
             } else {
                 $rawcolumns = [];
             }
-            foreach($rawcolumns as $c){
+            foreach ($rawcolumns as $c) {
                 array_push($columns, [
                     "name"          => $c,
                     "required"      => true,
-                    "label"         => str_replace('_',' ',strtoupper($c)),
-                    "align"         => $types[$c] === 'integer'?'right': 'left',
+                    "label"         => str_replace('_', ' ', strtoupper($c)),
+                    "align"         => $types[$c] === 'integer' ? 'right' : 'left',
                     "field"         => $c,
                     "type"          => "text",
                     "show"          => true,
@@ -218,62 +249,97 @@ class ReportController extends BaseController
                     "option_filter" => true
                 ]);
             }
-            
+
             $jsonFile = [];
             $jsonFile['query'] = $data['query'];
             $jsonFile['columns'] = $columns;
-            $trimmedName = strtolower(trim($data['name'])).'.json';
-            $filePath = resource_path("data/queries/reports/user/{$trimmedName}");
+            $trimmedNameJson = strtolower(trim($data['name'])) . '.json';
+            $trimmedNameSql = strtolower(trim($data['name'])) . '.sql';
+            $filePathJson = resource_path("data/queries/reports/user/{$trimmedNameJson}");
+            $filePathSql = resource_path("data/queries/reports/user/{$trimmedNameSql}");
             $dir = resource_path("data/queries/reports/user");
             if (!File::isDirectory($dir)) {
                 File::makeDirectory($dir, 0775, true, true);
-            }else {
+            } else {
                 File::makeDirectory($dir, 0775, true, true);
             }
 
-            file_put_contents($filePath, json_encode($jsonFile, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE));
-            chmod($filePath, 0664);
+            file_put_contents($filePathJson, json_encode($jsonFile, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE));
+            file_put_contents($filePathSql, $data['query']);
+            chmod($filePathJson, 0664);
+            chmod($filePathSql, 0664);
             return Response::ok('Query berhasil disimpan');
-        }catch(Exception $e){
+        } catch (Exception $e) {
             rollBack();
-            return $this->setAlert('error','Gagal',$e->getMessage());
+            return $this->setAlert('error', 'Gagal', $e->getMessage());
         }
     }
 
-    public function deleteQuery(Request $request, $name){
+    public function deleteQuery(Request $request, $name)
+    {
         $this->allowAccessModule('report.report', 'delete');
-        if(empty($name)) return Response::badRequest('Nama file tidak ditemukan');
-        $filePath = resource_path('data/queries/reports/user/'.$name);
-        if (File::exists($filePath)) {
-            File::delete($filePath);
-            return Response::ok('Laporan berhasil dihapus');
-        }
-        
+        if (empty($name)) return Response::badRequest('Nama file tidak ditemukan');
+        $filePath = resource_path('data/queries/reports/user/' . $name);
+        $filePathSql = resource_path('data/queries/reports/user/' . str_replace('.json', '.sql', $name));
+
+        if (File::exists($filePath)) File::delete($filePath);
+        if (File::exists($filePathSql)) File::delete($filePathSql);
+
+        return Response::ok('Laporan berhasil dihapus');
     }
 
-    private function getQueryFiles($folder = 'reports',$showNav = false){
+    public function download(Request $request)
+    {
+        $this->allowAccessModule('report.report', 'download');
+        if (empty($request->path)) return Response::badRequest('Path tidak ditemukan');
+        $setting = Setting::where('name', 'toko')->first();
+        $pathSql = trim(str_replace('.json', '.sql', $request->path));
+        $sql = $this->getResourceSql($pathSql);
+        $this->_query = $sql;
+
+        $model = (new DynamicModel())->setTable(DB::raw("({$this->_query}) as t"));
+
+        $data = $model->get();
+        // dd($data->toArray());
+        if ($data->isEmpty()) {
+            return Response::badRequest('Data kosong');
+        }
+
+        $meta = [
+            'title' => $request->name,
+            'Toko' => $setting->data->namaToko,
+            'Periode' => Carbon::now()->format('M-Y')
+        ];
+        return Excel::download(
+            new DynamicCollectionExport($data, $meta),
+            $request->name . '.xlsx'
+        );
+    }
+
+    private function getQueryFiles($folder = 'reports', $showNav = false)
+    {
         $files = [];
         $dirDefault = new Directory();
         $dirDefault->withNavigation = $showNav;
-        $default = $dirDefault->scan('data/queries/'.$folder.'/default')->get();
+        $default = $dirDefault->scan('data/queries/' . $folder . '/default')->where('extension', '!=', 'sql')->get();
 
         $dirUser = new Directory();
         $dirUser->withNavigation = $showNav;
-        $user = $dirUser->scan('data/queries/'.$folder.'/user')->get()->reject(function($item){
+        $user = $dirUser->scan('data/queries/' . $folder . '/user')->where('extension', '!=', 'sql')->get()->reject(function ($item) {
             return $item['name'] == '.gitignore';
         });
         foreach ($default as $k => $d) {
-            $exist = $user->where('name',$d['name'])->first();
-            if($exist) array_push($files,$exist);
-            else array_push($files,$d);
+            $exist = $user->where('name', $d['name'])->first();
+            if ($exist) array_push($files, $exist);
+            else array_push($files, $d);
         }
         foreach ($user as $k => $u) {
-            $exist = $default->where('name',$u['name'])->first();
-            if(empty($exist)) array_push($files,$u);
+            $exist = $default->where('name', $u['name'])->first();
+            if (empty($exist)) array_push($files, $u);
         }
 
-        $files = collect($files)->map(function($item){
-            $item['label'] = strtoupper(str_replace('_',' ',explode('.',$item['name'])[0]));
+        $files = collect($files)->map(function ($item) {
+            $item['label'] = strtoupper(str_replace('_', ' ', explode('.', $item['name'])[0]));
             return $item;
         });
 
@@ -282,28 +348,30 @@ class ReportController extends BaseController
 
     private function protectQuery($query)
     {
-        // Trim spasi dan newline
         $query = trim($query);
 
-        // Hapus komentar SQL tipe "-- ..." atau "/* ... */"
-        $query = preg_replace('/--.*(\n|$)/', '', $query);       // komentar baris tunggal
-        $query = preg_replace('/\/\*.*?\*\//s', '', $query);     // komentar blok
+        // hapus komentar
+        $query = preg_replace('/--.*(\n|$)/', '', $query);
+        $query = preg_replace('/\/\*.*?\*\//s', '', $query);
 
         $query = trim($query);
 
-        $upper = strtoupper($query);
+        // wajib SELECT
         if (!preg_match('/^SELECT\s+/i', $query)) {
             return false;
         }
 
-        $forbidden = ['UPDATE', 'DELETE', 'INSERT', 'DROP', 'ALTER', 'CREATE', 'TRUNCATE', 'REPLACE'];
-        foreach ($forbidden as $word) {
-            if (str_contains($upper, $word)) {
-                return false;
-            }
+        // forbidden SQL commands (full word only)
+        $forbiddenPattern = '/\b(UPDATE|DELETE|INSERT|DROP|ALTER|CREATE|TRUNCATE|REPLACE)\b/i';
+
+        if (preg_match($forbiddenPattern, $query)) {
+            return false;
+        }
+
+        if (str_contains($query, ';')) {
+            return false;
         }
 
         return true;
     }
-
 }
